requirements.txt
annotated-types
anyio
asgiref
certifi
charset-normalizer
colorama
distro
dj-database-url
Django

google-auth
google-genai
gunicorn
h11
httpcore
httpx
idna
lxml
packaging
psycopg2-binary
pyasn1
pyasn1_modules
pydantic
pydantic_core
python-docx
python-dotenv
requests
rsa
sniffio
sqlparse
tenacity
tqdm
typing-inspection
typing_extensions
tzdata
urllib3
websockets
whitenoise
